Completed the Midterm Review
I added all the elements from requirement three down. I fixed the byline to the specific specifications and fixed the fonts to what they were supposed to be. I fixed my images so they looked better on the page and created a block quote and colored it green. I finished by adding an unordered list with links to more stories.